# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/task/add": {
            "post": {
                "tags": [
                    "Task API"
                ],
                "summary": "Add new task",
                "operationId": "76f5b62d3d3d2c6dd92c9023ac5db925",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Complete user APIs"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Make user APIs CRUD..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Added Successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/task/{id}/detail": {
            "get": {
                "tags": [
                    "Task API"
                ],
                "summary": "Get detail task",
                "operationId": "b8daa208e9209fa7347af79c346ec672",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and get detail task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 1,
                                                "user_id": 3,
                                                "name": "Complete user APIs",
                                                "desc": "Make user APIs CRUD...",
                                                "created_at": "2023-08-29T04:33:26.000000Z",
                                                "updated_at": null
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/{id}/update": {
            "post": {
                "tags": [
                    "Task API"
                ],
                "summary": "Update this task",
                "operationId": "ccc97e276c4d9c0aa060a39f2967da99",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Complete user APIs"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Make user APIs CRUD...1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and update this task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Updated Successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/{id}/delete": {
            "get": {
                "tags": [
                    "Task API"
                ],
                "summary": "Delete task",
                "operationId": "2419f6c92bde7e1b6b954eaf7c987a07",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and delete this task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Deleted Successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}